開発試作/CRM-AI-Matching-PoC/02_テスト/catalyst-function/requirements.txt
zcatalyst-sdk
openai
pinecone
python-dotenv
requests
